+++ src/main/render/thrift-server/struct/decode.ts
@@ -630,13 +630,16 @@ function readEndForFieldType(fieldType: ContainerType): ts.CallExpression {
 export function readStructBegin(): ts.ExpressionStatement {
     return createMethodCallStatement(
         COMMON_IDENTIFIERS.input,
-        'readStructBegin',
+        COMMON_IDENTIFIERS.readStructBegin,
     )
 }
 
 // input.readStructEnd()
 export function readStructEnd(): ts.ExpressionStatement {
-    return createMethodCallStatement(COMMON_IDENTIFIERS.input, 'readStructEnd')
+    return createMethodCallStatement(
+        COMMON_IDENTIFIERS.input,
+        COMMON_IDENTIFIERS.readStructEnd,
+    )
 }
 
 // input.readFieldBegin()
