+++ src/main/sys.ts
@@ -2,6 +2,9 @@ import * as fs from 'fs'
 import * as os from 'os'
 import * as path from 'path'
 
+import { print } from './printer'
+import { IGeneratedFile } from './types'
+
 function rootDir(): string {
     if (os.platform() === 'win32') {
         return process.cwd().split(path.sep)[0]
@@ -41,4 +44,22 @@ export function mkdir(dirPath: string): void {
     } else if (parts.length > 0) {
         createPath(parts, process.cwd())
     }
+}
+
+export function saveFiles(files: Array<IGeneratedFile>, outDir: string): void {
+    files.forEach((next: IGeneratedFile) => {
+        const outPath: string = path.resolve(
+            outDir,
+            next.path,
+            `${next.name}.ts`,
+        )
+
+        mkdir(path.dirname(outPath))
+
+        try {
+            fs.writeFileSync(outPath, print(next.body, true))
+        } catch (err) {
+            throw new Error(`Unable to save generated files to: ${outPath}`)
+        }
+    })
 }
